requirements.txt
twitchio
python-decouple

